More website fixes
