+++ frontend/components/Adventure/index.css
@@ -4,10 +4,12 @@
 @import '../../styles/breakpoints.css';
 @import '../../styles/colors.css';
 
-.copy {
+.root {
   display: flex;
   flex-direction: column;
+}
 
+.root > * {
   @include Breakpoint-tabletAndAbove {
     width: $five-columns;
   }
@@ -19,10 +21,23 @@
   }
 }
 
+.root > *:not(:last-child) {
+  margin-bottom: $spacing-56;
+}
+
 .copy > *:not(:last-child) {
   padding-bottom: $spacing-28;
 }
 
 .title {
-  margin-bottom: $spacing-56;
+
+}
+
+.left {
+
+}
+
+.right {
+
+  margin-left: auto;
 }
