+++ frontend/components/Heading/index.js
@@ -41,10 +41,6 @@ export default class Heading extends React.Component {
 
     const HeadingComponent = `h${level}`;
 
-    return (
-      <div>
-        <HeadingComponent styleName={styles}>{children}</HeadingComponent>
-      </div>
-    );
+    return <HeadingComponent styleName={styles}>{children}</HeadingComponent>;
   }
 }
