+++ frontend/components/Plan/index.js
@@ -9,14 +9,16 @@ import './index.css';
 
 const Plan = () => (
   <Section color="red" id="Plan">
+    <div styleName="title">
+      <Heading color="white" weight="bold" underline>
+        O Plano
+      </Heading>
+    </div>
     <div styleName="copy">
-      <div styleName="title">
-        <Heading color="white" weight="bold" underline>
-          O Plano
-        </Heading>
-      </div>
       <div styleName="events">
-        {eventsList.map(event => <Event key={event.day} {...event} />)}
+        {eventsList.map(event => (
+          <Event key={event.day} {...event} />
+        ))}
       </div>
     </div>
   </Section>
